train_ui/requirements.txt
gradio
google-cloud-storage
google-cloud-build
google-cloud-aiplatform